feat: revamp the style of select component dialog
